Se agrego la funcion de registrar compras y se modifico la interfaz para que funcionara en thymeleaf

+++ src/main/java/com/scm/scm/Mapper/CompraMapper.java
@@ -20,7 +20,7 @@ public class CompraMapper implements RowMapper<Compra> {
         compra.setPrecio(rs.getDouble("Precio"));
         compra.setUnidad(rs.getString("Unidad"));
         compra.setEstado(rs.getString("Estado"));
-        compra.setTotal(rs.getDouble("Total"));
+        compra.setTotal(rs.getFloat("Total"));
         compra.setFecha(rs.getString("Fecha"));
 
         return compra;
